Disable local caching
This change removes the possibility to enable local caching in the calendar
settings and disables it by default.
Local caching of calendar events can be useful if the calender is stored
on a remote server (like caldav), but does not make much sense here, as the
events are generated from the locally stored contacts.

Enabled local caching also causes a bug since Lightning 4:
The age is no longer displayed on the recurrent events, because Lightning
defaults to local caching and does only cache the initial event.