Rework calendar creation for TB 68
Instead of providing our own page for selecting between local normal and
local address book calendars, we now just add the address book option
directly to the inital page. This reduces the complexity and works
around a bug where the content of our overlay wizard page is not
properly displayed.

We now also disable address books in the selection, for which there
already is an calendar, where we would previously silently fail.