Gracefully handle the next button of the calendar creation screen
When multiple extensions add an entry to the calendar-type radiogroup,
we have to be careful with handling the status of the advance button,
which is shared. To handle this, we use the fact that there are two
events fired when the type is changed: select and command

In the first event (select), we can update the button without caring
about other extensions, while in the second one (command), we only
update if we are currently selected. This allows resetting the state
in the first event and giving another extension the control over the
button in the second event.